add duration chart

+++ src/main/java/parkhouse/CarIF.java
@@ -4,7 +4,7 @@ public interface CarIF {
     int nr();
     long begin();
     long end();
-    int duration();
-    int price();
+    long duration();
+    double price();
     void updateParams(String[] params); // Update params, added by Dominik 26.04
 }
